backend/requirements.txt

fastapi
uvicorn
gunicorn
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
